+++ src/input_context/input_modifier/smooth_nudge.rs
@@ -47,6 +47,12 @@ impl InputModifier for SmoothNudge {
         }
 
         let target_value = value.as_axis3d();
+        if self.current_value.distance_squared(target_value) < 1e-4 {
+            // Snap to the target if the distance is too small.
+            self.current_value = target_value;
+            return value;
+        }
+
         self.current_value
             .smooth_nudge(&target_value, self.decay_rate, time.delta_secs());
 
@@ -89,5 +95,17 @@ mod tests {
             modifier.apply(&actions, &time, true.into()),
             0.55067104.into()
         );
+    }
+
+    #[test]
+    fn snapping() {
+        let mut modifier = SmoothNudge::default();
+        let actions = ActionsData::default();
+        let mut time = Time::default();
+        time.advance_by(Duration::from_millis(100));
+        modifier.current_value = Vec3::X * 0.99;
+        assert_eq!(modifier.apply(&actions, &time, 1.0.into()), 1.0.into());
+        modifier.current_value = Vec3::X * 0.98;
+        assert_ne!(modifier.apply(&actions, &time, 1.0.into()), 1.0.into());
     }
 }