+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Snap to the target in `SmoothNudge` when the current value is close enough.
+
 ## [0.6.0] - 2025-01-13
 
 ### Changed